sua lai phieu nhap hang

+++ VLXD_API/DTOs/PhieuNhapKho/ChiTietPhieuNhapCreateDto.cs
@@ -3,7 +3,7 @@
     public class ChiTietPhieuNhapCreateDto
     {
         public int MaSanPham { get; set; }
-        public double SoLuong { get; set; }
+        public decimal SoLuong { get; set; } // Khớp kiểu decimal(18,2) trong DB mới
         public decimal GiaNhap { get; set; }
     }
 }
